requirements.txt
arch==6.3.0
fastparquet==2024.2.0
lightgbm==4.6.0
matplotlib==3.10.8
numba==0.60.0
numpy==2.4.3
pandas==3.0.1
python_dateutil==2.9.0.post0
scikit_learn==1.8.0
scipy==1.17.1
scores==2.5.0
seaborn==0.13.2
statsmodels==0.14.4
torch==2.2.2
tqdm==4.66.5
xarray==2024.11.0
pyarrow==15.0.2
openpyxl==3.1.2